requirements.txt
PaddleOCR
paddlepaddle
Pillow

